+++ remappedSrc/dev/sidgames/wmo/Reference.java
@@ -1,0 +1,17 @@
+package dev.sidgames.wmo;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Reference {
+    public static final Logger LOGGER = LoggerFactory.getLogger(Reference.MOD_ID);
+
+    public static final String MOD_ID = "wmo";
+    public static final String MOD_NAME = "Way More Ores";
+    public static final String MOD_VERSION = "23w16a";
+    public static final String[] MOD_AUTHOR = {"SidGames5", "LeWolfYT"};
+
+    public static final String MINECRAFT_VERSION = "1.19";
+    public static final String LOADER_VERSION = "0.14.9";
+    public static final String YARN_MAPPINGS = "1.19+build.4";
+}