+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 
 ## Technical Updates
 
+- Updated Minecraft (1.19 -> 1.19.4)
+- Updated Yarn mappings (1.19+build.4 -> 1.19.4+build.2)
+- Updated loader version (0.14.9 -> 0.14.19)
+- Updated Fabric API (0.58.0+1.19 -> 0.78.0+1.19.4)
+
 # Overall changelog for 1.7.0
 
 ## Additions